requirements.txt
fpdf==1.7.2
markdownify==0.14.1
pypdf==5.1.0
pytesseract==0.3.13
pillow==11.0.0
